+++ internal/forgejo/client.go
@@ -24,7 +24,6 @@ type Client struct {
 }
 
 func NewTokenClient(opts vcs.NewTokenClientOptions) (vcs.Client, error) {
-	url := fmt.Sprintf("https://%s", opts.Hostname)
 	options := make([]forgejo.ClientOption, 0, 2)
 	options = append(options, forgejo.SetToken(opts.Token))
 	if opts.SkipTLSVerification {
@@ -37,7 +36,7 @@ func NewTokenClient(opts vcs.NewTokenClientOptions) (vcs.Client, error) {
 		}
 		options = append(options, forgejo.SetHTTPClient(client))
 	}
-	rv, err := forgejo.NewClient(url, options...)
+	rv, err := forgejo.NewClient(opts.BaseURL.String(), options...)
 	if err != nil {
 		return nil, err
 	}