+++ internal/gitlab/gitlab.go
@@ -2,14 +2,12 @@
 package gitlab
 
 import (
+	"github.com/leg100/otf/internal"
 	oauth2gitlab "golang.org/x/oauth2/gitlab"
 )
 
-const (
-	DefaultHostname string = "gitlab.com"
-)
-
 var (
-	OAuthEndpoint = oauth2gitlab.Endpoint
-	OAuthScopes   = []string{"read_user", "read_api"}
+	DefaultBaseURL = internal.MustWebURL("https://gitlab.com")
+	OAuthEndpoint  = oauth2gitlab.Endpoint
+	OAuthScopes    = []string{"read_user", "read_api"}
 )