+++ internal/gitlab/kind.go
@@ -3,24 +3,32 @@ package gitlab
 import (
 	"context"
 
+	"github.com/leg100/otf/internal"
 	"github.com/leg100/otf/internal/vcs"
 )
 
-func RegisterVCSKind(vcsService *vcs.Service, hostname string, skipTLSVerification bool) {
+const KindID vcs.KindID = "gitlab"
+
+func RegisterVCSKind(vcsService *vcs.Service, apiURL *internal.WebURL, skipTLSVerification bool) {
 	vcsService.RegisterKind(vcs.Kind{
-		ID:   vcs.KindID("gitlab"),
+		ID:   KindID,
 		Icon: Icon(),
 		TokenKind: &vcs.TokenKind{
-			Description: tokenDescription(hostname),
+			Description: tokenDescription(apiURL.Host),
 		},
-		Hostname:     hostname,
+		DefaultURL:   apiURL,
 		EventHandler: HandleEvent,
-		NewClient: func(ctx context.Context, cfg vcs.Config) (vcs.Client, error) {
+		NewClient: func(ctx context.Context, cfg vcs.ClientConfig) (vcs.Client, error) {
 			return NewTokenClient(vcs.NewTokenClientOptions{
 				Token:               *cfg.Token,
-				Hostname:            hostname,
+				BaseURL:             apiURL,
 				SkipTLSVerification: skipTLSVerification,
 			})
+		},
+		TFEServiceProviders: []vcs.TFEServiceProviderType{
+			vcs.ServiceProviderGitlab,
+			vcs.ServiceProviderGitlabCE,
+			vcs.ServiceProviderGitlabEE,
 		},
 	})
 }