+++ internal/integration/daemon_config.go
@@ -3,6 +3,7 @@ package integration
 import (
 	"time"
 
+	"github.com/leg100/otf/internal"
 	"github.com/leg100/otf/internal/authenticator"
 	"github.com/leg100/otf/internal/daemon"
 	"github.com/leg100/otf/internal/engine"
@@ -15,6 +16,8 @@ type config struct {
 	daemon.Config
 	// skip creation of default organization
 	skipDefaultOrganization bool
+	// skip setting up an automatic github server stub
+	skipGithubStub bool
 	// github stub server options
 	githubOptions []github.TestServerOption
 }
@@ -47,7 +50,10 @@ func withGithubOptions(opts ...github.TestServerOption) configOption {
 
 func withGithubHostname(hostname string) configOption {
 	return func(cfg *config) {
-		cfg.GithubHostname = hostname
+		cfg.GithubHostname = internal.MustWebURL(hostname)
+		// setting a hostname implies the test is setting up its own stub so
+		// skip setting up another stub
+		cfg.skipGithubStub = true
 	}
 }
 
@@ -124,5 +130,18 @@ func withServerRunnerDebug() configOption {
 func withServerRunnerSandbox() configOption {
 	return func(cfg *config) {
 		cfg.RunnerConfig.Sandbox = true
+	}
+}
+
+func withKeyPairPaths(private, public string) configOption {
+	return func(cfg *config) {
+		cfg.PrivateKeyPath = private
+		cfg.PublicKeyPath = public
+	}
+}
+
+func withHostname(hostname string) configOption {
+	return func(cfg *config) {
+		cfg.Host = hostname
 	}
 }