+++ internal/vcs/config.go
@@ -1,6 +1,0 @@
-package vcs
-
-type Config struct {
-	Token        *string
-	Installation *Installation
-}